+++ CHANGELOG.md
@@ -21,6 +21,9 @@ See [0Ver](https://0ver.org/).
 - Add `default_error` parameter to `returns.converters.maybe_to_result`,
   which provides a default error value for `Failure`
 
+### Bugfixes
+
+- Add pickling support for `UnwrapFailedError` exception
 
 ## 0.23.0